#25 [MOVE] CalendarInfoBox 컴포넌트로 분리
